# back-end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API Documentation",
        "contact": {
            "email": "info@yeagger.com"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://localhost:8000/",
            "description": "Learning env"
        }
    ],
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "userAuth"
                ],
                "summary": "Add New user",
                "operationId": "6e78f3cf2fc8947f5418f1d14849a151",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "user name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "user email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User created successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "userAuth"
                ],
                "summary": "log user",
                "operationId": "b9af18be56384e6dd7e58525ca77c038",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "user email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User created successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "userAuth"
                ],
                "summary": "logout user",
                "operationId": "ce51e2c7dbb4cb57dc1d359d430a6813",
                "responses": {
                    "200": {
                        "description": "User created successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/progress": {
            "post": {
                "tags": [
                    "Progress"
                ],
                "summary": "Add New progress",
                "operationId": "96c81d81ace26cea054aef770ee75f90",
                "parameters": [
                    {
                        "name": "performance",
                        "in": "query",
                        "description": "user performance",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "measurements",
                        "in": "query",
                        "description": "user measurements",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "weight",
                        "in": "query",
                        "description": "User Weight",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "User",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User created successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/showProgress": {
            "get": {
                "tags": [
                    "Progress"
                ],
                "summary": "Get Progress Details",
                "operationId": "1937b747b7663d01d9373a6a87203a20",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/progress/{id}": {
            "put": {
                "tags": [
                    "Progress"
                ],
                "summary": "Update progress status",
                "operationId": "72b0261355573bc0ece5085260d307e6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the progress to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Fields to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Newstatut": {
                                        "description": "New status for the progress",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "progress status updated successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Progress"
                ],
                "summary": "Delete Progress",
                "description": "Deletes a Progress by ID",
                "operationId": "deleteProgress",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the progress to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/api/progress/{id}/complete": {
            "put": {
                "tags": [
                    "Progress"
                ],
                "summary": "Update progress complete",
                "operationId": "a795ee2a957c4bc492321cacfd751552",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the progress to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Fields to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Newstatut": {
                                        "description": "New status for the progress",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "progress status updated successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/progress/history": {
            "get": {
                "tags": [
                    "Progress"
                ],
                "summary": "Get Progress history",
                "operationId": "c41220deebad4c595927e2729ff4d619",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}